+++ saga/font.cpp
@@ -125,18 +125,18 @@ int FONT_Shutdown(void)
 	return R_SUCCESS;
 }
 
-int FONT_Load(ulong font_rn, int font_id)
+int FONT_Load(uint32 font_rn, int font_id)
 {
 
 	R_FONT_HEADER fh;
 	R_FONT *font;
 	R_FONT_STYLE *normal_font;
 
-	uchar *fontres_p;
+	byte *fontres_p;
 	size_t fontres_len;
 	size_t remain;
 
-	const uchar *read_p;
+	const byte *read_p;
 
 	int nbits;
 	int c;
@@ -498,7 +498,7 @@ FONT_GetStringWidth(int font_id,
 	size_t ct;
 	int width = 0;
 	int ch;
-	const uchar *txt_p;
+	const byte *txt_p;
 
 	if (!FontModule.init) {
 		return R_FAILURE;
@@ -516,7 +516,7 @@ FONT_GetStringWidth(int font_id,
 	font = FontModule.fonts[font_id];
 	assert(font != NULL);
 
-	txt_p = (const uchar *) test_str;
+	txt_p = (const byte *) test_str;
 
 	for (ct = test_str_ct;
 	    *txt_p && (!test_str_ct || ct > 0); txt_p++, ct--) {
@@ -598,16 +598,16 @@ FONT_Out(R_FONT_STYLE * draw_font,
     size_t draw_str_ct, int text_x, int text_y, int color)
 {
 
-	const uchar *draw_str_p;
+	const byte *draw_str_p;
 
-	uchar *c_data_ptr;
+	byte *c_data_ptr;
 	int c_code;
 
 	int char_row;
 
-	uchar *output_ptr;
-	uchar *output_ptr_min;
-	uchar *output_ptr_max;
+	byte *output_ptr;
+	byte *output_ptr_min;
+	byte *output_ptr_max;
 
 	int row;
 	int row_limit;
@@ -623,7 +623,7 @@ FONT_Out(R_FONT_STYLE * draw_font,
 		return R_SUCCESS;
 	}
 
-	draw_str_p = (const uchar *) draw_str;
+	draw_str_p = (const byte *) draw_str;
 	ct = draw_str_ct;
 
 	/* Draw string one character at a time, maximum of 'draw_str'_ct 
@@ -700,7 +700,7 @@ FONT_Out(R_FONT_STYLE * draw_font,
 				    c_bit--) {
 
 					if ((*c_data_ptr >> c_bit) & 0x01) {
-						*output_ptr = (uchar) color;
+						*output_ptr = (byte) color;
 					}
 					output_ptr++;
